Update cli.py
I added the altitude 'filter'  in line 70 so that the camera was not constantly following the landing aircraft at the local airport.  I wanted to be able to identify the aircraft that were going over the top.  I live so close to the final approach at my local airport the camera was fixated on following these landing planes because they were always the closest aircraft, I'm literally less than 2 miles from the approach line.  I don't have the coding skills to add this as a option in the command line options - but I'm guessing it would be a simple addition, my attempt ended badly.  I chose 30000 feet as my boundary as this is supposedly the area where contrails are most likely to be started to be produced.